Disable virtual line mode
I used to have all motions work for "virtual" lines, e.g. lines as they
are displayed on screen, as opposed to lines as they are in the actual
file. This is very useful for text, but less so for code.

Now virtual line editing is only enabled in the text filetypes, namely
text and markdown.