Change mapping for editing alternate (last) buffer
I used to use double space for editing the alternate buffer, but this is
breaking down with terminal mode, which I'm using more and more. I now
use the less congestive C-w<space> instead, which works in all modes and
is still easy to type.

A warning is added for the old mapping to remind me of the change until
muscle memory is trained.